# docs/swagger.yaml
definitions:
  commons.SwaggerApiResponseError:
    properties:
      data: {}
      message:
        type: string
      success:
        default: false
        type: boolean
    type: object
  commons.SwaggerApiResponseSuccessWithData:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
  commons.SwaggerApiResponseSuccessWithoutData:
    properties:
      data: {}
      message:
        type: string
      success:
        default: true
        type: boolean
    type: object
...
  models.EmployeeRequest:
    properties:
      branch_id:
        type: integer
      name:
        type: string
    type: object
  models.ItemRequest:
    properties:
      item_type_id:
        type: integer
      name:
        type: string
      price:
        type: integer
      profit:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple store management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple Store Management API
  version: "1.0"
paths:
...
  /api/branch/employees/{id}:
    get:
      consumes:
      - application/json
      description: get branch with employees
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch With Employees
      tags:
      - branch
  /api/branch/items/{id}:
    get:
      consumes:
      - application/json
      description: get branch with Iitems
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Branch With Items
      tags:
      - branch
...
  /api/employee:
    get:
      consumes:
      - application/json
      description: get all employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get All Employees
      tags:
      - employee
    post:
      consumes:
      - application/json
      description: create employee
      parameters:
      - description: Employee Request
        in: body
        name: employeeRequest
        required: true
        schema:
          $ref: '#/definitions/models.EmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Create Employee
      tags:
      - employee
  /api/employee/{id}:
    delete:
      consumes:
      - application/json
      description: delete employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Delete Employee
      tags:
      - employee
    get:
      consumes:
      - application/json
      description: get employee by id
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Employee
      tags:
      - employee
    put:
      consumes:
      - application/json
      description: update employee
      parameters:
      - description: Employee Request
        in: body
        name: employeeRequest
        required: true
        schema:
          $ref: '#/definitions/models.EmployeeRequest'
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Update Employee
      tags:
      - employee
  /api/employee/top/{branchId}:
    get:
      consumes:
      - application/json
      description: get top employee
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Top Employee
      tags:
      - employee
  /api/item:
    get:
      consumes:
      - application/json
      description: get all items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get All Items
      tags:
      - item
    post:
      consumes:
      - application/json
      description: create a new item
      parameters:
      - description: Item Request
        in: body
        name: itemRequest
        required: true
        schema:
          $ref: '#/definitions/models.ItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Create Item
      tags:
      - item
  /api/item/{id}:
    delete:
      consumes:
      - application/json
      description: delete item by id
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Delete Item
      tags:
      - item
    get:
      consumes:
      - application/json
      description: get item by id
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Item
      tags:
      - item
    put:
      consumes:
      - application/json
      description: update item by id
      parameters:
      - description: Item Request
        in: body
        name: itemRequest
        required: true
        schema:
          $ref: '#/definitions/models.ItemRequest'
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithoutData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Update Item
      tags:
      - item
...
  /api/itemType/items/{id}:
    get:
      consumes:
      - application/json
      description: get items of item type
      parameters:
      - description: Item Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Items Of Item Type
      tags:
      - itemType
...
  /api/salesData/employee/{employeeId}:
    get:
      consumes:
      - application/json
      description: get sales data by employee
      parameters:
      - description: Employee ID
        in: path
        name: employeeId
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseSuccessWithData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.SwaggerApiResponseError'
      security:
      - Bearer: []
      summary: Get Sales Data Employee
      tags:
      - salesData
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"